[♻️refactor]: components/Icons 폴더를 Shared 폴더 내부로 이동
