+++ src/components/KanbanBoard/IssuePreviewList.tsx
@@ -3,7 +3,7 @@ import { css } from '@emotion/react';
 import { theme, ThemeType } from '@/styles/theme';
 import { RequirementStateType } from '@/types/requirement';
 
-import EllipsisIcon from '../icons/EllipsisIcon';
+import EllipsisIcon from '../Shared/Icons/EllipsisIcon';
 import RequirementCheckToggle from '../Requirement/RequirementCheckToggle';
 
 type IssuePreviewListProps = {
