+++ src/components/Shared/Icons/CloseIcon.tsx
