+++ src/components/Shared/Icons/CopyIcon.tsx
