+++ src/components/Shared/Icons/DeleteIcon.tsx
