+++ src/components/Shared/Icons/GPTIcon.tsx
