+++ src/components/gpt/GPTGenerateCodeBtn.tsx
@@ -11,7 +11,7 @@ import { useRequirement } from '@/hooks/useRequirement';
 import { RequirementStateType } from '@/types/requirement';
 import { generateSearchPrompt } from '@/utils/gpt';
 
-import GPTIcon from '../icons/GPTIcon';
+import GPTIcon from '../Shared/Icons/GPTIcon';
 
 type GPTGenerateCodeBtnProps = {
   requirement: RequirementStateType;
