+++ src/components/gpt/info/APIKeyForm.tsx
@@ -5,8 +5,8 @@ import { useRecoilState } from 'recoil';
 
 import { gptAtom } from '@/atoms/gptAtom';
 import HoverIcon from '@/components/Shared/HoverIcon';
-import CheckIcon from '@/components/icons/CheckIcon';
-import DeleteIcon from '@/components/icons/DeleteIcon';
+import CheckIcon from '@/components/Shared/Icons/CheckIcon';
+import DeleteIcon from '@/components/Shared/Icons/DeleteIcon';
 import useInput from '@/hooks/useInput';
 import { theme, ThemeType } from '@/styles/theme';
